Add endpoint to list videos with a default limit of 10
Introduced a new GET endpoint to retrieve a list of videos, with the results limited to 10 by default. This change simplifies access to the video data and improves the organization of the application's routes. Additional comments with example URLs were also included for better clarity and documentation of the API endpoints.